requirements.txt
discord.py[voice]>=2.4.0
yt-dlp>=2025.1.0
PyNaCl>=1.5.0
